[Refactor] Extract listen-for-network-messages.ts
